update: update directory composition
