+++ src/pages/_app.tsx
@@ -1,26 +1,25 @@
-import '../styles/globals.css'
-import type { AppProps } from 'next/app'
-import Amplify from 'aws-amplify'
-import config from '../src/aws-exports'
-Amplify.configure(config)
+import "@/styles/globals.css";
+import type { AppProps } from "next/app";
+import Amplify from "aws-amplify";
+import config from "../aws-exports";
+Amplify.configure(config);
 
 if (typeof window !== "undefined") {
   const { host } = window.location;
-  if (config.oauth.redirectSignIn.includes(',')) {
+  if (config.oauth.redirectSignIn.includes(",")) {
     const filterHost = (url: string) => new URL(url).host === host;
     config.oauth.redirectSignIn = config.oauth.redirectSignIn
-      .split(',')
+      .split(",")
       .filter(filterHost)
       .shift()!;
     config.oauth.redirectSignOut = config.oauth.redirectSignOut
-      .split(',')
+      .split(",")
       .filter(filterHost)
       .shift()!;
   }
 }
 
-
 function MyApp({ Component, pageProps }: AppProps) {
-  return <Component {...pageProps} />
+  return <Component {...pageProps} />;
 }
-export default MyApp
+export default MyApp;
