+++ src/pages/index.tsx
@@ -1,6 +1,6 @@
 import type { NextPage } from 'next';
 import Head from 'next/head';
-import styles from '../styles/Home.module.css';
+import styles from '@/styles/Home.module.css';
 import Link from 'next/link';
 
 const Home: NextPage = () => {
