+++ src/pages/ssg-demo.tsx
@@ -1,17 +1,17 @@
-import Head from 'next/head';
-import styles from '../styles/Home.module.css';
-import { GetStaticProps } from 'next';
-import Link from 'next/link';
+import Head from "next/head";
+import styles from "@/styles/Home.module.css";
+import { GetStaticProps } from "next";
+import Link from "next/link";
 
 type Props = { nowDate: string; pageTitle: string };
 
-export default function ISRDemo(props: Props) {
+export default function SSGDemo(props: Props) {
   return (
     <div className={styles.container}>
       <Head>
         <title>{props.pageTitle}</title>
-        <meta name='description' content='Generated by create next app' />
-        <link rel='icon' href='/favicon.ico' />
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
       </Head>
 
       <main className={styles.main}>
@@ -19,7 +19,7 @@ export default function ISRDemo(props: Props) {
         <p>What time is it now?</p>
         <p>It&apos;s {props.nowDate}.</p>
         <p className={styles.description}>
-          <Link href={'/'}>
+          <Link href={"/"}>
             <a>Go back</a>
           </Link>
         </p>
@@ -32,8 +32,7 @@ export const getStaticProps: GetStaticProps<Props> = async (context) => {
   return {
     props: {
       nowDate: new Date().toLocaleString(),
-      pageTitle: 'ISR Demo',
+      pageTitle: "SSG Demo",
     },
-    revalidate: 5, // ISR settings
   };
 };
