+++ src/pages/ssr-demo.tsx
@@ -1,17 +1,17 @@
-import Head from 'next/head';
-import styles from '../styles/Home.module.css';
-import { GetStaticProps } from 'next';
-import Link from 'next/link';
+import Head from "next/head";
+import styles from "@/styles/Home.module.css";
+import { GetServerSideProps } from "next";
+import Link from "next/link";
 
 type Props = { nowDate: string; pageTitle: string };
 
-export default function SSGDemo(props: Props) {
+export default function SSRDemo(props: Props) {
   return (
     <div className={styles.container}>
       <Head>
         <title>{props.pageTitle}</title>
-        <meta name='description' content='Generated by create next app' />
-        <link rel='icon' href='/favicon.ico' />
+        <meta name="description" content="Generated by create next app" />
+        <link rel="icon" href="/favicon.ico" />
       </Head>
 
       <main className={styles.main}>
@@ -19,7 +19,7 @@ export default function SSGDemo(props: Props) {
         <p>What time is it now?</p>
         <p>It&apos;s {props.nowDate}.</p>
         <p className={styles.description}>
-          <Link href={'/'}>
+          <Link href={"/"}>
             <a>Go back</a>
           </Link>
         </p>
@@ -28,11 +28,13 @@ export default function SSGDemo(props: Props) {
   );
 }
 
-export const getStaticProps: GetStaticProps<Props> = async (context) => {
+export const getServerSideProps: GetServerSideProps<Props> = async (
+  context
+) => {
   return {
     props: {
       nowDate: new Date().toLocaleString(),
-      pageTitle: 'SSG Demo',
+      pageTitle: "SSR Demo",
     },
   };
 };
